+++ besu/src/main/java/org/hyperledger/besu/cli/DefaultCommandValues.java
@@ -129,7 +129,7 @@ public interface DefaultCommandValues {
   String DEFAULT_PLUGINS_OPTION_NAME = "--plugins";
 
   /** The constant DEFAULT_HALT_ON_PLUGIN_ERROR_OPTION_NAME. */
-  String DEFAULT_HALT_ON_PLUGIN_ERROR_OPTION_NAME = "--halt-on-plugin-error";
+  String DEFAULT_HALT_ON_PLUGIN_ERROR_OPTION_NAME = "--plugin-continue-on-error";
 
   /** The constant DEFAULT_PLUGINS_EXTERNAL_ENABLED_OPTION_NAME. */
   String DEFAULT_PLUGINS_EXTERNAL_ENABLED_OPTION_NAME = "--Xplugins-external-enabled";